feat: 플레이어 피격 추가, 적 추가, 웨이브 추가 및 수정
- hp, maxhp를 float로 바꿈
- 적 스프라이트와 데이터를 추가, 그에 따른 프리팹 생성
- 새 적들을 웨이브에 배치

+++ Assets/Scripts/UI/HUD.cs
@@ -84,9 +84,9 @@ public class HUD : MonoBehaviour
         killCountText.text = $"{killCount}";
     }
 
-    private void UpdateHpMeter(int currentHp, int maxHp)
+    private void UpdateHpMeter(float currentHp, float maxHp)
     {
-        hpMeter.value = (float)currentHp / maxHp;
+        hpMeter.value = currentHp / maxHp;
     }
 
     private bool TryValidateSettings(out string error)
